+++ tests/framework/grid/GridViewTest.php
@@ -281,7 +281,7 @@ class GridViewTest extends TestCase
                 'options' => [
                     'dataProvider'   => new ArrayDataProvider(['allModels' => []]),
                     'id'             => 'test_grid_view',
-                    'filterSelector' => static fn($widgetId, $filterId) => "$widgetId foo $filterId bar",
+                    'filterSelector' => static fn ($widgetId, $filterId) => "$widgetId foo $filterId bar",
                 ]
             ]
         );
@@ -295,7 +295,7 @@ class GridViewTest extends TestCase
                 'options' => [
                     'dataProvider'           => new ArrayDataProvider(['allModels' => []]),
                     'id'                     => 'test_grid_view',
-                    'filterSelector'         => static fn($widgetId, $filterId) => "$widgetId foo $filterId bar",
+                    'filterSelector'         => static fn ($widgetId, $filterId) => "$widgetId foo $filterId bar",
                     'overrideFilterSelector' => true
                 ]
             ]